examples/langgraph/requirements.txt
langchain-mcp-adapters>=0.2.1
langgraph>=0.3.0
langchain-anthropic>=0.3.0
python-dotenv>=1.0.0
